requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-annoying==0.8.3
django-bcrypt==0.9.2
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-haystack==2.4.0
django-markdown-deux==1.0.5
django-pagination==1.0.7
django-registration-redux==1.2
django-storages==1.1.8
django-taggit==0.16.2
django-tastypie==0.12.2
django-templatetag-sugar==1.0
ecdsa==0.13
elasticsearch==1.6.0
Fabric==1.10.2
gunicorn==19.3.0
lxml==3.4.4
markdown2==2.3.0
mercurial==3.4
paramiko==1.15.2
parsedatetime==1.5
psycopg2==2.6.1
py-bcrypt==0.4
pycrypto==2.6.1
Pygments==2.0.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
python-postmark==0.4.7
raven==5.5.0
requests==2.7.0
six==1.9.0
smartypants==1.8.6
sqlparse==0.1.15
stripe==1.23.0
typogrify==2.0.7
urllib3==1.11
